+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## base-2.2.0, go-1.13-r20, go-1.14-r15, go-1.15-r6 (2020/11/08)
+
+* Add support for `linux/ppc64le` and `linux/s390x` archs (#18)
+
 ## go-1.14-r14, go-1.15-r5 (2020/11/06)
 
 * Go 1.14.11, 1.15.4